requirements.txt
django
factory_boy
djangorestframework
django-environ